refactor: Place all internal-use extension members in a single file
